# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_upload_documents_v1_cases__case_id__documents_post": {
        "properties": {
          "files": {
            "items": {
              "contentMediaType": "application/octet-stream",
              "type": "string"
            },
            "title": "Files",
            "type": "array"
          }
        },
        "required": [
          "files"
        ],
        "title": "Body_upload_documents_v1_cases__case_id__documents_post",
        "type": "object"
      },
      "CaseEnvelope": {
        "description": "``Case`` plus the API-shape-only ``_links`` placeholder.",
        "properties": {
          "_links": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Links",
            "type": "object"
          },
          "assigned_to_user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned To User Id"
          },
          "closure_date": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closure Date"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "customer_metadata": {
            "$ref": "#/components/schemas/CustomerMetadata"
          },
          "id": {
            "maxLength": 31,
            "minLength": 31,
            "pattern": "^case_[0-9A-HJKMNP-TV-Z]{26}$",
            "title": "Id",
            "type": "string"
          },
          "latest_decision": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CaseLatestDecision"
              },
              {
                "type": "null"
              }
            ]
          },
          "risk_band": {
            "anyOf": [
              {
                "enum": [
                  "low",
                  "medium_low",
                  "medium_high",
                  "high"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Band"
          },
          "state": {
            "$ref": "#/components/schemas/CaseState"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "state",
          "customer_metadata",
          "created_at",
          "updated_at"
        ],
        "title": "CaseEnvelope",
        "type": "object"
      },
...
      "CaseLatestDecision": {
        "description": "Per-case decision summary surfaced on ``GET /v1/cases/{case_id}``.\n\nStory 7.6 / AC #5. ``sealed_ledger_entry_id`` is ``None`` while the\ndecision is still in the 120-second undo window (state\n``pending_seal``); populated once the seal-time ledger entry is\nwritten. Story 7.7 owns the ``decisions`` table that backs this; the\nenvelope addition lands here so the cockpit-ui's\n``SealedIndicator`` can read the seal id without a second round-\ntrip.",
        "properties": {
          "decision_id": {
            "minLength": 1,
            "title": "Decision Id",
            "type": "string"
          },
          "outcome": {
            "enum": [
              "approve",
              "decline",
              "approve_with_conditions",
              "escalate_to_edd"
            ],
            "title": "Outcome",
            "type": "string"
          },
          "sealed_ledger_entry_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sealed Ledger Entry Id"
          }
        },
        "required": [
          "decision_id",
          "outcome"
        ],
        "title": "CaseLatestDecision",
        "type": "object"
      },
      "CaseListResponse": {
        "description": "Pagination envelope per ``architecture.md#Format Patterns``.",
        "properties": {
          "has_more": {
            "default": false,
            "title": "Has More",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/CaseEnvelope"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "CaseListResponse",
        "type": "object"
      },
      "CaseState": {
        "description": "Lifecycle states. See ``ALLOWED_TRANSITIONS`` for the edge set.",
        "enum": [
          "intake_scheduled",
          "decision_ready",
          "pending_seal",
          "committed",
          "escalated",
          "closed"
        ],
        "title": "CaseState",
        "type": "string"
      },
...
      "CustomerMetadata": {
        "description": "Customer-side facts attached to a case.\n\n``extra`` is an extensibility escape hatch so fixtures (Story 2-4) can carry\ndemo-only fields without schema churn. Production code should prefer named\nfields over ``extra`` once the bank-buyer scope revives.",
        "properties": {
          "country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country"
          },
          "customer_name": {
            "minLength": 1,
            "title": "Customer Name",
            "type": "string"
          },
          "customer_type": {
            "anyOf": [
              {
                "enum": [
                  "individual",
                  "company"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Type"
          },
          "extra": {
            "additionalProperties": true,
            "title": "Extra",
            "type": "object"
          }
        },
        "required": [
          "customer_name"
        ],
        "title": "CustomerMetadata",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ListResponse": {
        "description": "Reply to GET /documents.",
        "properties": {
          "case_id": {
            "title": "Case Id",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/StoredDocumentResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "case_id",
          "items"
        ],
        "title": "ListResponse",
        "type": "object"
      },
...
      "StoredDocumentResponse": {
        "description": "One document on disk, surfaced through the API.",
        "properties": {
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "size_bytes": {
            "minimum": 0.0,
            "title": "Size Bytes",
            "type": "integer"
          },
          "uploaded_at": {
            "format": "date-time",
            "title": "Uploaded At",
            "type": "string"
          }
        },
        "required": [
          "filename",
          "size_bytes",
          "uploaded_at"
        ],
        "title": "StoredDocumentResponse",
        "type": "object"
      },
...
      "UploadResponse": {
        "description": "Reply to a successful POST.",
        "properties": {
          "case_id": {
            "title": "Case Id",
            "type": "string"
          },
          "document_refs": {
            "items": {
              "type": "string"
            },
            "title": "Document Refs",
            "type": "array"
          },
          "uploaded": {
            "items": {
              "$ref": "#/components/schemas/StoredDocumentResponse"
            },
            "title": "Uploaded",
            "type": "array"
          }
        },
        "required": [
          "case_id",
          "uploaded",
          "document_refs"
        ],
        "title": "UploadResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Cockpit API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/cases": {
      "get": {
        "description": "Returns every case currently in the cockpit, newest first. Each item carries the case id, current state (intake_scheduled, decision_ready, etc.), customer metadata, and assigned officer. Used by the case_supervisor agent to answer 'what cases are there?'.",
        "operationId": "list_cases_v1_cases_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-Cockpit-Demo-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Cockpit-Demo-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List cases (newest first)",
        "tags": [
          "cases"
        ]
      }
    },
    "/v1/cases/{case_id}": {
      "get": {
        "operationId": "get_case_v1_cases__case_id__get",
        "parameters": [
          {
            "description": "Case ID (`case_<ULID>`)",
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "description": "Case ID (`case_<ULID>`)",
              "pattern": "^case_[0-9A-HJKMNP-TV-Z]{26}$",
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Cockpit-Demo-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Cockpit-Demo-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get a single case by ID",
        "tags": [
          "cases"
        ]
      }
    },
...
    "/v1/cases/{case_id}/documents": {
      "get": {
        "operationId": "list_case_documents_v1_cases__case_id__documents_get",
        "parameters": [
          {
            "description": "Case ID (`case_<ULID>`)",
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "description": "Case ID (`case_<ULID>`)",
              "pattern": "^case_[0-9A-HJKMNP-TV-Z]{26}$",
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Cockpit-Demo-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Cockpit-Demo-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List the case's uploaded documents",
        "tags": [
          "documents"
        ]
      },
      "post": {
        "description": "Accepts multipart/form-data with one or more files in the `files` field. Each file must be a valid PDF (magic-byte check) \u2264 10 MB. On success, returns the per-file metadata + the case's updated document_refs list.",
        "operationId": "upload_documents_v1_cases__case_id__documents_post",
        "parameters": [
          {
            "description": "Case ID (`case_<ULID>`)",
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "description": "Case ID (`case_<ULID>`)",
              "pattern": "^case_[0-9A-HJKMNP-TV-Z]{26}$",
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Cockpit-Demo-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Cockpit-Demo-User"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_documents_v1_cases__case_id__documents_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload one or more PDF documents to a case",
        "tags": [
          "documents"
        ]
      }
    },
    "/v1/cases/{case_id}/documents/{filename}": {
      "delete": {
        "operationId": "delete_case_document_v1_cases__case_id__documents__filename__delete",
        "parameters": [
          {
            "description": "Case ID (`case_<ULID>`)",
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "description": "Case ID (`case_<ULID>`)",
              "pattern": "^case_[0-9A-HJKMNP-TV-Z]{26}$",
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "filename",
            "required": true,
            "schema": {
              "title": "Filename",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Cockpit-Demo-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Cockpit-Demo-User"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete an uploaded document and its document_refs entry",
        "tags": [
          "documents"
        ]
      }
    },
    "/v1/cases/{case_id}/documents/{filename}/download": {
      "get": {
        "description": "Story 4 hardening \u2014 returns the case's PDF straight from ``./fixtures/uploads/<case_id>/<filename>`` so the analyst can preview the file the agents have been reading. 404 if the file isn't on disk; 400 on a malformed filename.",
        "operationId": "download_case_document_v1_cases__case_id__documents__filename__download_get",
        "parameters": [
          {
            "description": "Case ID (`case_<ULID>`)",
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "description": "Case ID (`case_<ULID>`)",
              "pattern": "^case_[0-9A-HJKMNP-TV-Z]{26}$",
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "filename",
            "required": true,
            "schema": {
              "title": "Filename",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download a stored document as application/pdf",
        "tags": [
          "documents"
        ]
      }
    },
...
  }
}